+++ openmetadata-ui/src/main/resources/ui/src/components/BotDetails/BotDetails.style.less
@@ -1,15 +1,3 @@
-.bot-details-buttons {
-  .ant-btn-icon-only {
-    width: 25px;
-    height: 25px;
-    padding: 0.4px 0;
-  }
-}
-.bot-details-card {
-  border: 1px rgb(221, 227, 234) solid;
-  border-radius: 4px;
-  box-shadow: 1px 1px 8px rgb(0 0 0 / 6%);
-}
 .display-name.ant-typography {
   margin-right: 0.5rem;
   margin-bottom: 0;
