+++ openmetadata-ui/src/main/resources/ui/src/components/BotDetails/DisplayNameComponent/DisplayNameComponent.interface.ts
@@ -1,0 +1,9 @@
+export interface DisplayNameComponentProps {
+  isDisplayNameEdit: boolean;
+  displayName: string | undefined;
+  onDisplayNameChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  setIsDisplayNameEdit: (value: boolean) => void;
+  handleDisplayNameChange: () => void;
+  displayNamePermission: boolean;
+  editAllPermission: boolean;
+}
