+++ openmetadata-ui/src/main/resources/ui/src/components/BotDetails/DisplayNameComponent/DisplayNameComponent.style.less
@@ -1,0 +1,5 @@
+@import url('../../../styles/variables.less');
+
+.add-display-name {
+  color: @text-grey-muted;
+}
