+++ openmetadata-ui/src/main/resources/ui/src/components/Users/Users.component.tsx
@@ -351,7 +351,7 @@ const Users = ({
   const getDescriptionComponent = () => {
     if (isAdminUser || isLoggedinUser || isAuthDisabled) {
       return (
-        <div className="tw-flex tw-items-center tw-justify-between">
+        <div className="flex items-center justify-between">
           <Description
             description={userData.description || ''}
             entityName={getEntityName(userData as unknown as EntityReference)}
