+++ openmetadata-ui/src/main/resources/ui/src/components/common/InheritedRolesCard/InheritedRolesCard.interface.ts
@@ -1,0 +1,5 @@
+import { User } from '../../../generated/entity/teams/user';
+
+export interface InheritedRolesCardProps {
+  userData: User;
+}
