+++ openmetadata-ui/src/main/resources/ui/src/components/common/InheritedRolesCard/InheritedRolesCard.style.less
@@ -1,0 +1,10 @@
+@import url('../../../styles/variables.less');
+
+.inherited-roles-heading {
+  font-weight: 500;
+  color: @text-color-secondary;
+}
+
+.inherited-no-description {
+  color: @text-grey-muted;
+}
