+++ openmetadata-ui/src/main/resources/ui/src/components/common/RolesCard/RolesCard.component.tsx
@@ -1,0 +1,155 @@
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { Button, Card, Select, Space } from 'antd';
+import { isArray, isNil, toLower } from 'lodash';
+import React, { useState } from 'react';
+import { useTranslation } from 'react-i18next';
+import { useAuthContext } from '../../../authentication/auth-provider/AuthProvider';
+import { TERM_ADMIN } from '../../../constants/constants';
+import { useAuth } from '../../../hooks/authHooks';
+import { getEntityName } from '../../../utils/CommonUtils';
+import SVGIcons from '../../../utils/SvgUtils';
+import RolesElement from '../RolesElement/RolesElement.component';
+import { RolesComponentProps } from './RolesCard.interfaces';
+
+const RolesCard = ({
+  roles,
+  userData,
+  updateUserDetails,
+  selectedRoles,
+  setSelectedRoles,
+}: RolesComponentProps) => {
+  const [isRolesEdit, setIsRolesEdit] = useState(false);
+
+  const { t } = useTranslation();
+  const { isAdminUser } = useAuth();
+  const { isAuthDisabled } = useAuthContext();
+
+  const handleRolesChange = () => {
+    // filter out the roles , and exclude the admin one
+    const updatedRoles = isArray(selectedRoles)
+      ? selectedRoles.filter((roleId) => roleId !== toLower(TERM_ADMIN))
+      : [];
+    // get the admin role and send it as boolean value `isAdmin=Boolean(isAdmin)
+    const isAdmin = isArray(selectedRoles)
+      ? selectedRoles.find((roleId) => roleId === toLower(TERM_ADMIN))
+      : [];
+
+    updateUserDetails({
+      roles: updatedRoles.map((roleId) => {
+        const role = roles.find((r) => r.id === roleId);
+
+        return { id: roleId, type: 'role', name: role?.name || '' };
+      }),
+      isAdmin: Boolean(isAdmin),
+    });
+
+    setIsRolesEdit(false);
+  };
+
+  const handleOnRolesChange = (selectedOptions: string[]) => {
+    if (isNil(selectedOptions)) {
+      setSelectedRoles([]);
+    } else {
+      setSelectedRoles(selectedOptions);
+    }
+  };
+
+  const userRolesOption = isArray(roles)
+    ? roles.map((role) => ({
+        label: getEntityName(role),
+        value: role.id,
+      }))
+    : [];
+
+  if (!userData.isAdmin) {
+    userRolesOption.push({
+      label: TERM_ADMIN,
+      value: toLower(TERM_ADMIN),
+    });
+  }
+
+  if (isAdminUser && !isAuthDisabled) {
+    return (
+      <Card
+        className="ant-card-feed relative page-layout-v1-left-panel mt-2.5"
+        key="roles-card"
+        title={
+          <div className="flex items-center justify-between">
+            <h6 className="mb-0">{t('label.roles')}</h6>
+            {!isRolesEdit && (
+              <button
+                className="m-l-xs focus:tw-outline-none tw-self-baseline"
+                data-testid="edit-roles"
+                onClick={() => setIsRolesEdit(true)}>
+                <SVGIcons
+                  alt="edit"
+                  className="m-b-xss"
+                  icon="icon-edit"
+                  title="Edit"
+                  width="16px"
+                />
+              </button>
+            )}
+          </div>
+        }>
+        <div className="mb-4">
+          {isRolesEdit ? (
+            <Space className="w-full" direction="vertical">
+              <Select
+                aria-label="Select roles"
+                className="w-full"
+                defaultValue={selectedRoles}
+                id="select-role"
+                mode="multiple"
+                options={userRolesOption}
+                placeholder={`${t('label.roles')}...`}
+                onChange={handleOnRolesChange}
+              />
+              <div className="flex justify-end" data-testid="buttons">
+                <Button
+                  className="text-sm mr-1"
+                  data-testid="cancel-roles"
+                  icon={
+                    <FontAwesomeIcon className="w-3.5 h-3.5" icon="times" />
+                  }
+                  size="small"
+                  type="primary"
+                  onMouseDown={() => setIsRolesEdit(false)}
+                />
+                <Button
+                  className="text-sm"
+                  data-testid="save-roles"
+                  icon={
+                    <FontAwesomeIcon className="w-3.5 h-3.5" icon="check" />
+                  }
+                  size="small"
+                  type="primary"
+                  onClick={handleRolesChange}
+                />
+              </div>
+            </Space>
+          ) : (
+            <RolesElement userData={userData} />
+          )}
+        </div>
+      </Card>
+    );
+  } else {
+    return (
+      <Card
+        className="relative page-layout-v1-left-panel mt-2.5"
+        key="roles-card"
+        title={
+          <div className="flex items-center justify-between">
+            <h6 className="mb-0">{t('label.roles')}</h6>
+          </div>
+        }>
+        <div className="flex items-center justify-between mb-4">
+          <RolesElement userData={userData} />
+        </div>
+      </Card>
+    );
+  }
+};
+
+export default RolesCard;
