+++ openmetadata-ui/src/main/resources/ui/src/components/common/RolesElement/RolesElement.styles.less
@@ -1,0 +1,5 @@
+@import url('../../../styles/variables.less');
+
+.roles-no-description {
+  color: @text-grey-muted;
+}
