+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/en-us.json
@@ -155,15 +155,14 @@
     "token-expiration": "Token Expiration",
     "select-auth-mechanism": "Select Auth Mechanism",
     "auth-mechanism": "Auth Mechanism",
-    "tokenEndpoint": "TokenEndpoint",
-    "clientId": "ClientId",
-    "secretKey": "SecretKey",
+    "token-end-point": "TokenEndpoint",
+    "client-id": "ClientId",
+    "secret-key": "SecretKey",
     "scopes": "Scopes",
     "okta-email": "Okta Service account Email",
-    "orgUrl": "OrgUrl",
-    "orgUrl-required": "OrgURL is required",
+    "org-url": "OrgUrl",
     "authority": "Authority",
-    "clientSecret": "ClientSecret",
+    "client-secret": "ClientSecret",
     "audience": "Audience",
     "enter-test-case-name": "Enter test case name",
     "test-type": "Test Type",
@@ -311,24 +310,24 @@
     "detail-plural": "Details",
     "ingestion": "Ingestion",
     "add-workflow-ingestion": "Add {{workflow}} Ingestion",
-    "clearAll": "Clear All",
+    "clear-all": "Clear All",
     "view-less": "View less",
     "view-more": "View more",
     "column-plural": "Columns",
     "add-display-name": "Add display name",
     "token-security": "Token Security",
-    "displayName": "displayName"
+    "teams": "Teams"
   },
   "message": {
     "service-email-required": "Service account Email is required",
     "token-expiration-required": "Token Expiration is required",
     "domain-required": "Domain is required",
-    "clientSecret-required": "ClientSecret is required",
-    "clientId-required": "ClientId is required",
+    "client-secret-required": "ClientSecret is required",
+    "client-id-required": "ClientId is required",
     "authority-required": "Authority is required",
     "scopes-required": "Scopes is required",
-    "secretKey-required": "SecretKey is required",
-    "tokenEndpoint-required": "TokenEndpoint is required",
+    "secret-key-required": "SecretKey is required",
+    "token-end-point-required": "TokenEndpoint is required",
     "privateKey-required": "PrivateKey is required",
     "auth-mechanism-required": "Auth Mechanism is required",
     "kill-successfully": "Successfully killed running workflows for",
@@ -373,7 +372,8 @@
     "no-permission-to-view": "You do not have the necessary permissions to view this data.",
     "no-schema-data-available": " No schema data available",
     "no-data-available-for-selected-filter": "No data found. Try changing the filters.",
-    "token-security-description": "Anyone who has your JWT Token will be able to send REST API requests to the OpenMetadata Server. Do not expose the JWT Token in your application code. Do not share it on GitHub or anywhere else online."
+    "token-security-description": "Anyone who has your JWT Token will be able to send REST API requests to the OpenMetadata Server. Do not expose the JWT Token in your application code. Do not share it on GitHub or anywhere else online.",
+    "org-url-required": "OrgURL is required"
   },
   "server": {
     "no-followed-entities": "You have not followed anything yet.",
