+++ openmetadata-ui/src/main/resources/ui/src/pages/BotDetailsPage/BotDetailsPage.tsx
@@ -17,7 +17,6 @@ import { compare } from 'fast-json-patch';
 import React, { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router-dom';
-import { useAuthContext } from '../../authentication/auth-provider/AuthProvider';
 import {
   getBotByName,
   getUserByName,
@@ -38,7 +37,6 @@ import { UserDetails } from '../../components/Users/Users.interface';
 import { NO_PERMISSION_TO_VIEW } from '../../constants/HelperTextUtil';
 import { Bot } from '../../generated/entity/bot';
 import { User } from '../../generated/entity/teams/user';
-import { useAuth } from '../../hooks/authHooks';
 import jsonData from '../../jsons/en';
 import { DEFAULT_ENTITY_PERMISSION } from '../../utils/PermissionsUtils';
 import { showErrorToast } from '../../utils/ToastUtils';
@@ -54,10 +52,8 @@ const BotDetailsPage = () => {
     DEFAULT_ENTITY_PERMISSION
   );
 
-  const { isAdminUser } = useAuth();
-  const { isAuthDisabled } = useAuthContext();
-
   const { t } = useTranslation();
+
   const fetchBotPermission = async (entityFqn: string) => {
     setIsLoading(true);
     try {
@@ -158,8 +154,6 @@ const BotDetailsPage = () => {
           botData={botData}
           botPermission={botPermission}
           botUserData={botUserData}
-          isAdminUser={Boolean(isAdminUser)}
-          isAuthDisabled={Boolean(isAuthDisabled)}
           revokeTokenHandler={revokeBotsToken}
           updateBotsDetails={updateBotsDetails}
           updateUserDetails={updateUserDetails}
