+++ openmetadata-ui/src/main/resources/ui/src/styles/app.less
@@ -341,11 +341,3 @@
   flex-direction: column;
   padding: 24px 16px 0px 16px;
 }
-
-.no-description {
-  color: @text-grey-muted;
-}
-.heading {
-  font-weight: 500;
-  color: @text-color-secondary;
-}
